Deal with missing screen file

+++ ReadScreen.py
@@ -35,15 +35,20 @@ class MyTable(QTableWidget):
     def open_sheet(self, Screen):
         app_path=os.path.abspath(os.path.dirname(sys.argv[0]))
         path=Path(app_path).joinpath("Screen_Database", ScreenFile[Screen])
-        with open(path, newline='') as csv_file:
-            self.setRowCount(0); self.setColumnCount(10)
-            my_screen = csv.reader(csv_file, delimiter=',', quotechar='"')
-            for row_data in my_screen:
-                row = self.rowCount()
-                self.insertRow(row)
-                if len(row_data) > 10:
-                    self.setColumnCount(len(row_data))
-                for column, stuff in enumerate(row_data):
-                    item = QTableWidgetItem(stuff)
-                    self.setItem(row, column, item)
-        del app_path, path, my_screen
+        
+        if Path(path).is_file():
+            with open(path, newline='') as csv_file:
+                my_screen = csv.reader(csv_file, delimiter=',', quotechar='"')
+                self.setRowCount(0); self.setColumnCount(10)
+                for row_data in my_screen:
+                    row = self.rowCount()
+                    self.insertRow(row)
+                    if len(row_data) > 10:
+                        self.setColumnCount(len(row_data))
+                    for column, stuff in enumerate(row_data):
+                        item = QTableWidgetItem(stuff)
+                        self.setItem(row, column, item)
+            del my_screen
+        else : return False
+                           
+        del app_path, path
